+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The following sections are considered for each release: **Added, Changed, Fixed, Security, Deprecated, Removed**
 
 ## [Unreleased]
+- Upgrade `ecl-builder` to `0.3.0`
+- Add `@aehrc/ecl-editor-react` dependency (required by `ecl-builder@0.3.0`)
 - Fix issue where the ticket that a concept was authored on was not correctly picking up bulk-brand/packsize updates
 - Update My task display logic to fall back to matching by username when email is not found
 - Configuration added for the automatic creation of synonyms on the product authoring screen if the term contains a substring. These can be created a new page which is reachable through the drop down on the top right.